+++ src/services/pathfinding/heap-item.ts
@@ -1,0 +1,4 @@
+export interface IHeapItem<T> {
+    heapIndex: number;
+    compareTo(item: T): number;
+}